refactor: 💡 navigation barを外だし
