+++ src/pages/home/containers/HomeContainer.vue
@@ -1,7 +1,8 @@
 <template>
   <v-app>
     <NavigationHeaderPage :drawer="drawer" @emit="toggleDraw" />
-    <HomeComponent :draw="drawer" @emit="toggleDraw" />
+    <NavigationBarPage :drawer="drawer" @emit="toggleDraw" />
+    <HomeComponent />
     <Footer />
   </v-app>
 </template>
@@ -10,6 +11,7 @@
 import { HomeComponent } from '../components';
 import { FooterPage } from '../../footer';
 import { NavigationHeaderPage } from '../../navigation-header';
+import { NavigationBarPage } from '../../navigation-bar';
 import { Component, Vue } from 'vue-property-decorator';
 
 @Component({
@@ -17,6 +19,7 @@ import { Component, Vue } from 'vue-property-decorator';
     HomeComponent,
     FooterPage,
     NavigationHeaderPage,
+    NavigationBarPage,
   },
 })
 export default class HomeContainer extends Vue {
