+++ src/pages/navigation-bar/components/index.ts
@@ -1,0 +1,1 @@
+export { default as NavigationBarComponent } from './NavigationBarComponent.vue';
