+++ src/pages/navigation-bar/containers/index.ts
@@ -1,0 +1,1 @@
+export { default as NavigationBarPage } from './NavigationBarContainer.vue';
